only materialize specs that we at least might need
this takes the dependency-api style approach to specs, and ignores the
known set of specs to instead only instantiate the ones that are in the
gemfile or are dependencies of some version of those gems. we could do
even better by returning an object that knows how to lazily return specs
for just gems that it's asked for, but Bundler can't use that yet.

the downside to this approach is that it means we have to eagerly fetch
the info files for each gem in the tree of possibilities. in the worst
case, where we have nothing cached yet, that can mean a lot of requests.

on the upside, in subsequent installs, when most of the gem info is
already local, this is _way_ faster than even the dependency api.